requirements.txt
streamlit
cohere
openai-whisper
audio-recorder-streamlit
python-dotenv
pydub